Writing Data Class
Data Class can be accessed from all sub classes.
We write every data into that Class and then SendData to Pc or Mando from this Class. 
Also when we get new data (OnDataReceived) we save it here.
then we only save at the end this data Class to a file.